backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymongo==4.10.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
groq==0.13.1
tavily-python==0.5.0
python-dotenv==1.0.1
websockets==14.1
pydantic[email]==2.10.4
requests==2.32.3
reportlab==4.2.5
google-generativeai==0.8.3
httpx==0.28.1
